+++ seta-ui/seta_flask_server/blueprints/communities/models/models_dto.py
@@ -1,4 +1,3 @@
-from seta_flask_server.infrastructure.constants import (CommunityStatusConstants, RequestStatusConstants)
 from flask_restx import Model, fields
 
 user_info_model = Model("User Info",
