+++ seta-ui/seta_flask_server/blueprints/communities/resource_change_request.py
@@ -12,69 +12,15 @@ from seta_flask_server.infrastructure.decorators import auth_validator
 from seta_flask_server.infrastructure.scope_constants import SystemScopeConstants, ResourceScopeConstants, CommunityScopeConstants
 from seta_flask_server.infrastructure.constants import UserRoleConstants, ResourceRequestFieldConstants
 
-from .models.resource_request_dto import(new_change_request_parser, update_change_request_parser, change_request_model, user_info_model)
+from .models.resource_request_dto import(new_change_request_parser, change_request_model, user_info_model)
 
 resource_change_request_ns = Namespace('Resource Change Requests', validate=True, description='SETA Resource Change Requests')
 resource_change_request_ns.models[user_info_model.name] = user_info_model
 resource_change_request_ns.models[change_request_model.name] = change_request_model
 
-@resource_change_request_ns.route('/change-requests/pending', methods=['GET', 'POST'])
-class ResourceChangeRequestList(Resource):
-    '''Get list of all pending change requests for resources'''
-    
-    @inject
-    def __init__(self, usersBroker: IUsersBroker, 
-                 changeRequestsBroker: IResourceChangeRequestsBroker, 
-                 api=None, *args, **kwargs):
-        self.usersBroker = usersBroker
-        self.changeRequestsBroker = changeRequestsBroker
-        
-        super().__init__(api, *args, **kwargs)
-        
-    @resource_change_request_ns.doc(description='Retrieve all pending change requests for resources.',
-        responses={
-                    int(HTTPStatus.OK): "'Retrieved pending change requests.",
-                    int(HTTPStatus.FORBIDDEN): "Insufficient rights, scope 'resource/change_request/approve' required"
-                },
-        
-        security='CSRF')
-    @resource_change_request_ns.marshal_list_with(change_request_model, mask="*", skip_none=True)
-    @auth_validator()
-    def get(self):
-        '''
-        Retrieve all pending change requests for resources, available to sysadmins
-
-        Permissions: either scope "/seta/resource/change_request/approve" or "Administrator" role
-        '''
-        
-        identity = get_jwt_identity()
-        auth_id = identity["user_id"]
-        
-        #verify scope
-        user = self.usersBroker.get_user_by_id(auth_id)        
-        if user is None or user.is_not_active():
-            abort(HTTPStatus.FORBIDDEN, "Insufficient rights.")
-
-        hasApproveRight = user.role.lower() == UserRoleConstants.Admin.lower() or user.has_system_scope(SystemScopeConstants.ApproveResourceChangeRequest)        
-        if not hasApproveRight:        
-            abort(HTTPStatus.FORBIDDEN, "Insufficient rights.")
-
-        requests = self.changeRequestsBroker.get_all_pending()
-        for request in requests:
-            requested_by = self.usersBroker.get_user_by_id(request.requested_by, load_scopes=False)
-            if requested_by:
-                request.requested_by_info = requested_by.user_info
-
-            if request.reviewed_by:
-                reviewed_by = self.usersBroker.get_user_by_id(request.reviewed_by, load_scopes=False)
-                if reviewed_by:
-                    request.reviewed_by_info = reviewed_by.user_info
-
-        return requests
-
 @resource_change_request_ns.route('/<string:resource_id>/change-requests', methods=['GET','POST'])
 @resource_change_request_ns.param("resource_id", "Resource identifier") 
-class ResourceCreateChangeRequest(Resource):
+class ResourceChangeRequests(Resource):
     """Handles HTTP POST requests to URL: /resources/{resource_id}/change-requests."""
     
     @inject
@@ -196,7 +142,7 @@ class ResourceCreateChangeRequest(Resource):
         return requests
         
 
-@resource_change_request_ns.route('/<string:resource_id>/change-requests/<string:request_id>', methods=['GET', 'PUT'])
+@resource_change_request_ns.route('/<string:resource_id>/change-requests/<string:request_id>', methods=['GET'])
 @resource_change_request_ns.param("resource_id", "Resource identifier") 
 @resource_change_request_ns.param("request_id", "Change request identifier")
 class ResourceChangeRequest(Resource):
@@ -253,57 +199,3 @@ class ResourceChangeRequest(Resource):
                 request.reviewed_by_info = reviewed_by.user_info
         
         return request
-    
-    @resource_change_request_ns.doc(description='Approve/reject request',        
-    responses={
-                int(HTTPStatus.OK): "Request updated.", 
-                int(HTTPStatus.FORBIDDEN): "Insufficient rights, scope 'resource/change_request/approve' required",
-                int(HTTPStatus.NOT_FOUND): "Request not found."
-                },
-    security='CSRF')
-    @resource_change_request_ns.expect(update_change_request_parser)
-    @auth_validator()
-    def put(self, resource_id, request_id):
-        '''
-        Update a change request for resource, available to sysadmins
-
-        Permissions: scope '/seta/resource/change_request/approve' or "Administrator" role
-        '''
-        
-        identity = get_jwt_identity()
-        auth_id = identity["user_id"]
-        
-        user = self.usersBroker.get_user_by_id(auth_id)
-
-        if user is None or user.is_not_active():
-            abort(HTTPStatus.FORBIDDEN, "Insufficient rights.")
-
-        hasApproveRight = user.role.lower() == UserRoleConstants.Admin.lower() or user.has_system_scope(SystemScopeConstants.ApproveResourceChangeRequest)
-        if not hasApproveRight:
-            abort(HTTPStatus.FORBIDDEN, "Insufficient rights.")
-
-        request = None
-        request_dict = update_change_request_parser.parse_args()
-        status = request_dict["status"]
-        
-        try:  
-            request = self.changeRequestsBroker.get_request(resource_id=resource_id, request_id=request_id)
-            
-            if request is not None:              
-                request.status = status
-                request.reviewed_by = auth_id
-                
-                self.changeRequestsBroker.update(request)                
-        except:
-            app.logger.exception("Request->put")
-            abort(HTTPStatus.INTERNAL_SERVER_ERROR)   
-           
-        if request is None:
-            abort(HTTPStatus.NOT_FOUND)
-        
-        message = f"Request {status}."
-        response = jsonify(status="success", message=message)
-        response.status_code = HTTPStatus.OK
-        
-        return response    
-
