+++ seta-ui/seta_flask_server/factory.py
@@ -146,6 +146,7 @@ def register_blueprints(app):
     from .blueprints.auth_github import auth_github
     from .blueprints.profile import profile_bp_v1
     from .blueprints.catalogue import catalogue_bp_v1
+    from .blueprints.admin import admin_bp
     
     API_ROOT="/seta-ui/api"
                     
@@ -158,7 +159,9 @@ def register_blueprints(app):
     app.register_blueprint(communities_bp_v1, url_prefix=f"{API_ROOT}/v1")
 
     app.register_blueprint(catalogue_bp_v1, url_prefix=f"{API_ROOT}/v1")
-    
+
+    app.register_blueprint(admin_bp, url_prefix=f"{API_ROOT}")
+        
 def register_extensions(app):    
     github.init_app(app)
     scheduler.init_app(app)
